# Conflicts:
#   back/src/main/java/com/udteam/miristock/config/SecurityConfig.java

+++ back/src/main/java/com/udteam/miristock/dto/LimitPriceOrderDto.java
@@ -15,8 +15,8 @@ public class LimitPriceOrderDto {
     private String stockCode;
     private String stockName;
     private Integer memberNo;
-    private Integer limitPriceOrderPrice;
-    private Integer limitPriceOrderAmount;
+    private Long limitPriceOrderPrice;
+    private Long limitPriceOrderAmount;
     private Deal limitPriceOrderType;
 
     @Builder
